docs: support AMD (Rocm Kernel) - Merge upstream changes and update AMD tutorial (#741)
[Done]
1. Merged the latest upstream changes.
2. Split `docs/amd_tutorial/amd_build_dockerfile_page.rst` into three
parts and merged them into installation, quick start, and multi-node
training, respectively.
3. Can I still keep `amd_build_dockerfile_page.rst under
docs/amd_tutorial` (just left this in here will not indepently show a
page in the official docs) so that AMD cluster users can more easily
refer to it in one document, instead of having to find the settings
across different pages in the official docs?

---------

Co-authored-by: HL <linhaibin.eric@gmail.com>