[mcore] resharding model weights by per tensor (#1107)
## Motivation
This is an optimization approach using a per-tensor method to reduce the
additional memory required for model weights during the resharding
phase. Our ultimate goal is to enable mcore to have a method that aligns
with the `full_tensor()` function in FSDP and to deprecate the
`AllGatherPPModel` class in future versions. Currently, this task may
need to be broken down into several subtasks:

## Impact Analysis
1. The model accuracy has been tested on Qwen-7B in vllm version 0.8.2,
and it aligns with the accuracy of the previous method.
2. In terms of memory usage, the `pp_cache` in `AllGatherPPModel` has
been completely deprecated.
3. In terms of runtime, the performance is comparable to the original
method.

## TODO
- [x] Deprecate the `AllGatherPPModel` class in version 0.8.2.
- [x] Ensure forward compatibility for this method.
- [x] Completely deprecate the `AllGatherPPModel` class.