[recipe] sppo: SPPO algorithm implementation (#1222)
Here is a version of the SPPO algorithm implementation.

You can find more information about SPPO here:
[https://github.com/uclaml/SPPO/tree/main](https://github.com/uclaml/SPPO/tree/main)

In short, the main differences between SPPO and PPO are:

1. There is no need to use a critic model.
2. SoftMean is used as the AdvantageEstimator in the trainer.
3. Different loss functions.

I have made an attempt to implement minimal modifications without
altering the code outside the recipe. However, due to the following two
issues, the current code is not entirely elegant:


1. To modify the loss function (including both the loss itself and the
parameters passed in), it is sufficient to modify the `update_policy` in
the `DataParallelSPPOActor`. I attempted to patch this `update_policy`,
but it was unsuccessful. Therefore, I created a class that inherits from
`DataParallelSPPOActor` to override the `update_policy`.
2. Since `ActorRolloutRefWorker` imports `DataParallelSPPOActor` in the
`init_model` function, it also needs to inherit from
`ActorRolloutRefWorker` to override `init_model`. However, I encountered
an issue during inheritance. The base class of `verl`’s `single
controller` calls `actor_rolloutrefworker.super().__init__()`, and the
original `super()` in `ActorRolloutRefWorker` is `Worker`. If we
inherit, it would become `ActorRolloutRefWorker`, which requires passing
parameters to `super().__init__()`, but the `single controller base`
code does not provide any parameters, making inheritance impossible.

I have now submitted a draft PR and would appreciate any suggestions on
code modifications or optimizations!

---------

Co-authored-by: zhaochenyang20 <zhaochen20@outlook.com>