[Bug] Make `free_cache_engine` option workable in latest vLLM/SGLang
It looks like `actor_rollout_ref.rollout.free_cache_engine` control
option only works for vLLM version 0.5.4 and 0.6.3, the sleep / wake
up mode in vLLM engine, as well as release / resume memory occupation
in SGLang is enabled by default and there's no way to turn them off.

While always alllowing inference engine to free cache can be ideal,
it's unfortunately not supported on some devices, such as AMD Mi250x,
since it doesn't support virtual memory management:
https://github.com/vllm-project/vllm/pull/12695#issuecomment-2633919751

So we would need to be able to turn it off so that verl can run on
those devices.

In addition, it looks like we no longer need to enforce eager in latest
vLLM when we choose to free cache, so this PR also lifted this
restriction.

Signed-off-by: Hollow Man <hollowman@opensuse.org>