[vllm] feat: add `engine_kwargs`  in vllm_rollout_spmd to set params in vllm engine (#1442)
### Checklist Before Starting

- [x] Search for similar PR(s).

### What does this PR do?

> add `engine_kwargs` in vllm_rollout_spmd to set swap_space in vllm
engine.

### Specific Changes

> add `engine_kwargs` in vllm_rollout_spmd, which can be set in config
file. Same changes has been made in vllm_rollout. As the version of vllm
is update to 0.8, the default vllm_rollout worker becomes
vllm_rollout_spmd, which does not have `engine_kwargs` as in
vllm_rollout, so this RP want to complete it.

### Usage Example

> users can set vllm engine param such as `swap_space`, `seed` through
the `engine_kwargs` in config file. For example, if one want to set the
swap_space=32 in vllm, he can set the item in config like this
```bash
actor_rollout_ref.rollout.engine_kwargs.swap_space=32
```