[training_utils] feat: Support `assert_case` for sandbox fusion
It can be useful to support `assert_case` in addition to `inputs`
and `outputs`, which is seen in some datasets such as:
https://huggingface.co/datasets/a-m-team/AM-Thinking-v1-RL-Dataset

The `assert_case` field in `ground_truth` receives a list of strings,
which can be used to append additional test cases code after the
rollout generated code and provides more flexibility than pure
inputs and outputs checks.

Here, the design is that, `assert_case` can be used together with
`inputs` and `outputs`. In this case, all the three lists should
have the same length, so that, for each test case, we can append
additional check code while still checking the inputs and outputs.
If only `assert_case` is provided, we simply set `inputs` to a list
of empty strings and `outputs` to a list of `None` type, and we
set that when the expected output is None, we don't compare the
output of the generated code.

Signed-off-by: Hollow Man <hollowman@opensuse.org>