[worker] fix: use `is_multi_modal` field to check if it's a VLM
Address 2nd issue in https://github.com/volcengine/verl/pull/3281#issuecomment-3239570745

Currently, if we use ulysses sp, we rely on `multi_modal_inputs` to
check if it's a multi-modal model, but this can go wrong when we
set `data.return_multi_modal_inputs=False`, as that field won't
exist even if it's the VLM model.

As a result, we add an additional `is_multi_modal` field that
checks if the multi-modal key exists in the dataset, to see if
it's a multi-modal model or not.

Signed-off-by: Hollow Man <hollowman@opensuse.org>