[worker] fix: get all `multi_modal_inputs` keys with in a microbatch
Currently, the code gets the keys from the first row within the microbatch,
This can go wrong if the dataset is a mixture of pure-text with multi-modal,
where the first data in the microbatch is a pure-text one (no `pixel_values`
or `image_grid_thw` exists in the key), and the microbatch still contains
multi-modal data.

This PR fixes this issue by collecting all available keys for `multi_modal_inputs`
within the microbatch, and so that we can concatenate those multi-modal
tensors together without ignoring some of them under the above situation.

Signed-off-by: Hollow Man <hollowman@opensuse.org>