[worker] fix: support for vllm V0 deprecation version
Related to:
- https://github.com/vllm-project/vllm/pull/25901
- https://github.com/vllm-project/vllm/pull/25345

Now we first try to import `WorkerWrapperBase` from
`vllm.worker.worker_base`, if we have an error, we
append `v1` there.

For `compute_logits` patch, we can just remove the
import of `SamplingMetadata`, create a wrapper that
accepts any arguments with *args, **kwargs, and pass
them through to the original method, so that it can
be more flexible and future-proof.

Signed-off-by: Hollow Man <hollowman@opensuse.org>