[megatron, model] fix: VLMs using mbridge together with fused kernels
Current code is too rigid in checking support of fused forward and
will go wrong if we use mbridge and even it's a Qwen2_5VLModel,
since then the defined Qwen2.5VL multi-modal model class will be from
class definition in mbridge instead of the one in verl.

Also, many other VLMs supported in mbridge uses the `language_model`
attribute, and we just need to ensure that `model.language_model`
is an instance of mcore defined `GPTModel`, which should be a more
flexible and applicable way for checking support.

Signed-off-by: Hollow Man <hollowman@opensuse.org>