[trainer] fix: batch size mismatch with n>1 when gen_max for ReMax
We should not repeat directly on top of `gen_batch`, as
in `gen_max`, we need the original `gen_batch` so that
we can disable `do_sample` for rollout to calculate
the reward baseline.

```log
  File "verl/trainer/main_ppo.py", line 317, in run
    trainer.fit()
  File "verl/trainer/ppo/ray_trainer.py", line 1065, in fit
    batch = batch.union(gen_baseline_output)
  File "verl/protocol.py", line 802, in union
    self.batch = union_tensor_dict(self.batch, other.batch)
  File "verl/protocol.py", line 110, in union_tensor_dict
    assert tensor_dict1.batch_size == tensor_dict2.batch_size, (
AssertionError: Two tensor dict must have identical batch size. Got torch.Size([4096]) and torch.Size([16384])
```

Signed-off-by: Hollow Man <hollowman@opensuse.org>