[recipe] fix: DAPO using KL in reward
Currently, the DAPO recipe will fail if we set
`use_kl_in_reward` to True. This is because we
need to calculate `response_mask`, `old_log_prob`,
and/or `ref_log_prob` before we call `apply_kl_penalty`
to rewards.

Although KL divergence is removed according to DAPO
paper, it would still be neccessary to get this fixed,
as `apply_kl_penalty` is already there in the code,
also it can be good to support this to allow
exploration in non-standard ways.

Signed-off-by: Hollow Man <hollowman@opensuse.org>