[recipe, rollout] fix: Make one step off policy support more card combination
Because `ActorRolloutRefWorker` will call `initialize_model_parallel`
when init, any worker of subclass of `ActorRolloutRefWorker` needs to
meet the parallel configuration of the actor during initialization,
which leads to the one-step-off-policy rollout worker also having to
meet the corresponding card number to actor worker.

So if you have 32 cards for actor with PP=4 and EP=8, and 16 cards for
rollout with TP=4, you will meet `RuntimeError: decoder world_size (16)
is not divisible by expert_tensor_model_pipeline_parallel size (32)` in
megatron.

This patch deleted the megatron related codes for rollout worker.